LLMs/1-RAG_QA/requirement.txt
gradio==4.39.0
gradio-pdf==0.0.11
PyMuPDF==1.24.9
llama-index==0.10.58
llama-index-embeddings-huggingface==0.2.2
python-doctr==0.8.1
requests==2.28.2